+++ src/client/js/components/Header/Header.js
@@ -1,5 +1,5 @@
 import Component from "../../core/Component";
-import { createExtendsRelation } from "../../core/oop-utils";
+import { createExtendsRelation } from "../../oop-utils";
 import Search from "./Search/Search";
 
 function Header(...params) {
