+++ src/client/js/components/Header/Search/SearchUI/SearchCategoryList.js
@@ -1,5 +1,5 @@
 import Component from "../../../../core/Component";
-import { createExtendsRelation } from "../../../../core/oop-utils";
+import { createExtendsRelation } from "../../../../oop-utils";
 
 function SearchCategoryList(...params) {
   Component.call(this, ...params);
